+++ Unresize/unresize.cpp
@@ -74,7 +74,7 @@ size_t Unresize::tmp_size(PixelType type) const
 
 	// Line buffer for horizontal pass.
 	if (m_src_width != m_dst_width)
-		size += m_dst_width * 8;
+		size += (size_t)m_dst_width * 8;
 
 	return size;
 }
